+++ packages/dartcv/CHANGELOG.md
@@ -3,6 +3,9 @@
 * add `VideoCapture.grabAsync`
 * add `Mat.fromBuffer` to allow creating `Mat` from self-managed buffer
 * add `copyData` parameter to `Mat.fromVec`
+* remove `Mat.fromPtr`, merge it to `Mat.fromMat`
+* add `Mat.atU8`, `Mat.setU8`, etc.
+* remove deprecated `Mat.copyToWithMask`, `List<Mat>.asVecMat()`
 
 ## 1.1.2